packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-openclaw"
version = "0.1.0"
readme = "../../README.md"
license = {text = "AGPL-3.0"}
requires-python = ">=3.11"
dependencies = [
    "minions-sdk>=0.2.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[tool.hatch.build.targets.wheel]
packages = ["minions_openclaw"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
